Connecting to DB

+++ src/index.ts
@@ -1,5 +1,7 @@
 import {LogLevel, SapphireClient} from '@sapphire/framework';
 import {Intents} from 'discord.js';
+import {validateEnv} from './utils/validateEnv';
+import {connectDatabase} from './database/connectDatabase';
 
 require('dotenv').config({path: '../.env'});
 
@@ -13,10 +15,13 @@ const client = new SapphireClient({
 });
 
 const main = async () => {
+  if (!validateEnv()) return;
+
   try {
     client.logger.info('Bot is logging in');
     await client.login(process.env.BOT_TOKEN);
     client.logger.info('Bot has logged in');
+    await connectDatabase();
   } catch (error) {
     client.logger.fatal(error);
     client.destroy();
